경욱/정환

+++ src/main/java/com/example/dayte/reply/controller/PostReplyController.java
@@ -37,7 +37,7 @@ public class PostReplyController {
                                                      @AuthenticationPrincipal UserSecurityDTO principal) {
         postReplyDTO.setUser(modelMapper.map(principal, User.class));
 
-        int id = postReplyDTO.getId();
+        Long id = postReplyDTO.getId();
         Post post = postService.getPost(id);
         postReplyDTO.setPost(post);
         PostReply postReply = modelMapper.map(postReplyDTO, PostReply.class);
